Update README.md
This program, written in Python 3, finds products on Supreme's website in the 'new' category. Once a product is found it will automatically open the product page on your browser. Then the user can add to cart and checkout manually.  
Given the recent success of manual users on Supreme's website this program will allow you to navigate to the product page as quickly as possible. Even if the site is lagging you will be going straight to the product page which is a slight advantage over the normal manual user who navigates from the category page or even the shop home page.  
When you start the program you'll be prompted to enter a keyword. Once you hit return the program will search for the keyword (in an actual drop enter the keyword whenever and then hit return right at 11). It can be a single word or a phrase. For example in week 19 if you use the word 'Nike' (this program is not case sensitive, it doesn't matter what case you use) the page for the Nike Vapor Jet gloves will open. You can also use a small phrase such as 'camp cap' or 'reflective small' for the side panel camp cap or the reflective small box logo zip up, respectively. 
If the product you're searching for is not found (possibly due to a late drop or site lag) the program will search for the term every half a second ten more times before finally ending.  If your keyword is found in multiple products (for example, polartec in week 19) all the products will be pulled and an individual tab will be opened for each product with that keyword.   
Please note that this program has never been tested during an actual drop. I put it together after F/W18 ended. Also, if you open your default web browser and open gmail in a tab the new tab will populate in that same browser window and gmail will still be open too, this may give you an additional advantage with captcha's at checkout. I look forward to testing this during S/S19.  Good luck!